[fix] 신고 중복 데이터 filter
